Moving new crypto utility functions to _alpha while we consider them.
